Move RichTransactionInfo to types
